# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Toyow Workflow API
  version: 1.0.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /workflows/{id}:
    get:
      summary: Get workflow with versions
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Workflow detail }
    put:
      summary: Create next workflow version
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete workflow
      responses:
        "204": { description: Deleted }
...
  /runs/{workflowId}/start:
    post:
      summary: Start run for workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201": { description: Run created }
  /runs/{runId}:
    get:
      summary: Get run and step state
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Run detail }
...